requirements.txt
python==3.9.13
einops==0.4.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
patool==1.12
reformer-pytorch==1.4.4
pillow==9.4.0
scikit-learn==1.4.2
scipy==1.10.1
sktime==0.16.1
sympy==1.11.1
torch==2.0.0
optuna==3.4.0
torchinfo==1.7.2
torchvision==0.15.1
tqdm==4.64.1
fastai==2.7.14
